refactor: move sponsored-swap store out of delegation domain
