+++ src/__swaps__/screens/Swap/providers/SyncSwapStateAndSharedValues.tsx
@@ -7,7 +7,6 @@ import { create } from 'zustand';
 
 import { type ExtendedAnimatedAssetWithColors } from '@/__swaps__/types/assets';
 import { analytics } from '@/analytics';
-import { useIsSponsoredSwap } from '@/features/delegation/sponsoredSwapStore';
 import { type GasSettings } from '@/features/gas/hooks/useCustomGas';
 import { useSelectedGas } from '@/features/gas/hooks/useSelectedGas';
 import { useBackendNetworksStore } from '@/features/network/stores/backendNetworksStore';
@@ -29,6 +28,7 @@ import {
 import Routes from '@/navigation/routesNames';
 import { useUserAssetsStore } from '@/state/assets/userAssets';
 import { useStoreSharedValue } from '@/state/internal/hooks/useStoreSharedValue';
+import { useIsSponsoredSwap } from '@/state/swaps/sponsoredSwapStore';
 import { useSwapsStore } from '@/state/swaps/swapsStore';
 import { getUniqueId } from '@/utils/ethereumUtils';
 import { deepEqual } from '@/worklets/comparisons';
